subiendo registro de usuarios empleados

+++ backend/Financio/src/main/java/pe/edu/utp/Financio/Service/CajaService.java
@@ -1,11 +1,14 @@
 package pe.edu.utp.Financio.Service;
 import pe.edu.utp.Financio.entity.Caja;
+
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface CajaService {
     Caja guardarFondo(Caja caja);
-    Optional<Caja> guardarCierre(int idCaja, double cierre);
+    Optional<Caja> guardarCierre(int idCaja, BigDecimal cierre);
+    
     boolean eliminar(int idCaja);
     List<Caja> listar();
 }
